requirements.txt
Django
django-taggit
django-suit
django-mptt